# docs/swagger.yaml
basePath: /api
definitions:
...
  api.CreateOrderItemRequest:
    properties:
      color:
        type: string
      product_id:
        type: string
      quantity:
        type: integer
      size:
        type: string
    type: object
  api.CreateOrderRequest:
    properties:
      customer_notes:
        type: string
      items:
        items:
          $ref: '#/definitions/api.CreateOrderItemRequest'
        type: array
      payment_method:
        type: string
      promo_code:
        type: string
      shipping_info:
        $ref: '#/definitions/interfaces.ShippingInfo'
    type: object
...
  api.OrderResponse:
    properties:
      order:
        $ref: '#/definitions/interfaces.Order'
      order_id:
        type: string
      success:
        type: boolean
    type: object
  api.OrdersResponse:
    properties:
      has_more:
        type: boolean
      limit:
        type: integer
      offset:
        type: integer
      orders:
        items:
          $ref: '#/definitions/interfaces.Order'
        type: array
      total_count:
        type: integer
    type: object
...
  api.UpdateOrderStatusRequest:
    properties:
      payment_status:
        type: string
      status:
        type: string
      tracking_number:
        type: string
    type: object
...
  interfaces.JSONMap:
    additionalProperties: true
    type: object
  interfaces.Order:
    properties:
      cancelled_at:
        type: string
      created_at:
        type: string
      currency:
        type: string
      customer_notes:
        type: string
      delivered_at:
        type: string
      discount_amount:
        type: number
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/interfaces.OrderItem'
        type: array
      metadata:
        $ref: '#/definitions/interfaces.JSONMap'
      order_number:
        type: string
      payment_method:
        type: string
      payment_status:
        type: string
      promo_code:
        type: string
      shipped_at:
        type: string
      shipping_fee:
        type: number
      shipping_info:
        $ref: '#/definitions/interfaces.ShippingInfo'
      status:
        type: string
      subtotal:
        description: Pricing
        type: number
      total_price:
        type: number
      tracking_number:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  interfaces.OrderItem:
    properties:
      color:
        type: string
      created_at:
        type: string
      id:
        type: string
      line_total:
        type: number
      order_id:
        type: string
      product_id:
        type: string
      product_image:
        type: string
      product_name:
        description: Snapshot product data
        type: string
      product_slug:
        type: string
      quantity:
        type: integer
      size:
        type: string
      unit_price:
        type: number
    type: object
...
  interfaces.ShippingInfo:
    properties:
      address:
        type: string
      email:
        type: string
      firstName:
        type: string
      gouvernorat:
        type: string
      lastName:
        type: string
      notes:
        type: string
      phone:
        type: string
      postalCode:
        type: string
    type: object
...
host: localhost:5001
info:
  contact:
    email: malek.radhouen@gmail.com
    name: API Support
    url: https://www.magic.fr/support
  description: Here is our solution documentation and testing portal of provided functionalities
    to interact with our hypervisor tool.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://www.magic.com/terms/
  title: Magic - Inventory solution API
  version: "1.0"
paths:
...
  /admin/orders:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.OrdersResponse'
      summary: List all orders (admin)
      tags:
      - orders
  /admin/orders/{id}:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Status update
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/api.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/interfaces.Order'
      summary: Update order status
      tags:
      - orders
...
  /me/orders:
    get:
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.OrdersResponse'
      summary: List my orders
      tags:
      - orders
  /orders:
    post:
      consumes:
      - application/json
      parameters:
      - description: Order
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/api.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.OrderResponse'
      summary: Create order
      tags:
      - orders
  /orders/{id}:
    get:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Phone (required for guest orders)
        in: query
        name: phone
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/interfaces.Order'
      summary: Get order
      tags:
      - orders
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...